Create PrintSchema.scala
The schema lists all the columns in the DataFrame, along with their data types. In this case, all the columns are of type string. The nullable = true indicates that the columns allow null values.

This information is helpful for understanding the structure and types of the data in the DataFrame.